requirements.txt
click==8.1.6
contourpy==1.1.0
cycler==0.11.0
fonttools==4.41.1
joblib==1.3.1
kiwisolver==1.4.4
matplotlib==3.7.2
nltk==3.8.1
nose2==0.13.0
numpy==1.25.1
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
regex==2023.6.3
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
sklearn==0.0.post7
threadpoolctl==3.2.0
tqdm==4.65.0
tzdata==2023.3
